Add multiple image and video scraping
This now properly pulls down images and video for tweets, including
multiple images if there are such things. We have to fall back to V1.1
of the Twitter API to get video, but it's pretty damn quick, so not
really an issue.

Note: this could break if there's more than 900 tweets per archive per
day, due to rate limiting.

This also doesn't, yet, handle retweets and such.