Refactor video variant extraction
This refactors the video variant extraction a bit by checking explictly
for a `bitrate` key in each one. If there isn't one, we skip it because
that means it's probably a stream not a straight video file, which we
don't want.

This also reduces an extra loop which, let's be honest, means basically
nothing, but I'm a stickler to keeping that O(n) going.

Closes #2